More quick maths
Fixe dmost of the logic behind the halfsplit code. Issue now is that we can pinpoint the specific byte that Defender flags on, but since the next byte is technically okay, we increase by 50% and start the search again. This creates a dumb loop, which I'm sure I can just exit out of on condition.